Added base functionality in a chris object
Added on, off, addCss, removeCss and grab functionality in chris
object. Plan on moving the object over to a function like jquery
that can be called with $c. Off will remove all listeners if the
method is called without a second param and will remove all if
called with a function. All methods(except log) can be chained
together in a jquery like fashion.

TODO: Add README and maintain it with each commit.
Add append, prepend, clear, remove and other functions.
create list in README of functions to add to it as needed.